Deleting old 5-line Chinese image file
We can try this for Project 4, to test flexibility with wide range of input file sizes. But right now the code is optimized for somewhat bigger images, and the input files need to be roughly the same size so that font size and line spacing that is now set will work nicely.