To be deleted; see commit notes
This has Alex's newest set of tests. I copy-pasted them to the "consolidated" ipynb file. That file is newer than this (it has a lot of fixes). This file should be deleted but for now I wanted to keep it for the record (along with other similar copies in BackupBranch).